requirements.txt
streamlit
rdkit-pypi
pandas
